cast to size_t

+++ src/sentry_json.c
@@ -221,7 +221,8 @@ sentry__jsonwriter_write_double(sentry_jsonwriter_t *jw, double val)
     if (can_write_item(jw)) {
         char buf[50];
         // The MAX_SAFE_INTEGER is 9007199254740991, which has 16 digits
-        int written = sentry__snprintf_c(buf, sizeof(buf), "%.16g", val);
+        size_t written
+            = (size_t)sentry__snprintf_c(buf, sizeof(buf), "%.16g", val);
         // print `null` if we have printf issues or a non-finite double, which
         // can't be represented in JSON.
         if (written < 0 || written >= sizeof(buf) || !isfinite(val)) {
